requirements.txt
streamlit==1.13.0
opencv-python==4.6.0.66
numpy
tensorflow-hub
pillow==9.0.1
tensorflow==2.10.0
matplotlib==3.5.1
